+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # [Main]
 
 ### Fixed
-- Fix measuring large titles in modals.
+- Fixed list measurements with container headers.
 
 ### Added
 - Item reordering is now possible when using VoiceOver.